Fix removal of mediaelement when native controls are used
If native controls are used on iPhone, iPad or Android devices,
t.container is never set, thus it can not be used. However the video
player title isn't built in that case and does not need to be removed.

Without this fix, calling .remove() on the mediaelement leads to an
error message on iPad, iPhone and Android if the native controls are
used.