Less frame-greedy parentElement calculation for responsiveMode
This handles issue #2200. Before every iframe that existed above
t.container was instantly picked not regarding if another element was
visible. So the calculation of the element width failed even if it would
be possible.
With this change the code will first try to find a visible element
within the frame. Only if that is not possible (the node with the tag
name html is found and it is not visible), the frame element is used.