Bugfix: Make setPoster work on mobile devices with native controls
This fixes #2418.
If an mobile device is used and mediaelement is configured to use native
controls in this case, t.container is not defined. Therefore the
setPoster() method threw an error.
In that case now the poster attribute is used, which is the native way
of showing the poster image.